move path parser to separate class
